Fix note to self on fresh install
On account data initialization, a recipients entry is created
for the sender account that doesn't include the e164 (msisdn).

Subsequent calls to RecipientsTable::get either include the e164,
or the uuid, but never both, hence the e164 is never added to
the existing uuid record.

This patch aims to rectify this (in a slightly clumsy way).

See #235