Provide a way to send and receive attachments without temporary files
So far only implemented for attachments on outbound messages. Usage:

1. send a v2alpha1 binary_upload request with a the size (in bytes)
of the file. I used signaldctl's new `raw` subcommand for this:

  signaldctl raw v2alpha1 binary_upload '{"size": 286271}

it will return a transfer_id, used in the following steps.

2. send the file to the socket at /run/user/1000/signald/<transfer_id>.sock
I did this with netcat:

  nc -U /run/user/1000/signald/3b95bcf3-55d1-4664-a762-7e2cce259787.sock < /path/to/file

3. send a v2alpha1 send_message request. The format may be a
little unusual as this is only one of the enhancements in the v2
protocol:

  signaldctl raw v2alpha1 send_message '{"account": "<local account uuid>", "conversation": {"user": "<remote uuid>"}, "attachments": [{"content_type": "image/png", "transfer_id": "<transfer_id>"}]}'

Once the send_message request has been sent, the transfer will
occur, but signald will not close the binary transfer socket when
complete yet, so you must manually close the socket by pressing
ctrl-c on netcat. Once that happens, the send_message request
will get a response and the file will be uploaded.

I'm seeking feedback, this is just a first draft and likely to get
improved upon before being merged. Eventually I would like all
files in and out of signald to at least be available via something
like this.