Prevent deadlock in MessageReceiver.broadcast.
Exceptions in message encoding can prevent unlocking the reply lock.
This is not immediately apparent or an issue.  The reply lock is
reentrant.  So as long as messages are submitted from the
MessageReceiver thread, the reply lock can be locked and unlocked
seamlessly.

However, if another thread, like the one running the
SignalWebSockeHealhMonitor initiates a MessageReceiver.broadcast(),
that thread is unable to obtain the reply lock.

Worse, the MessageReceiver.broadcast() synchronizes the
SocketManagers's `listeners` list.  That in turn blocks access to
`sockets.size()` in MessageReceiver.run().  At that point, signald
stops receiving message from the websocket connection to the server.